// src/openapi-client/apis/DtsResourceApi.ts
export interface DtsListPostRequest {
    dtsFilter?: DtsFilter;
}