// src/openapi-client/models/DtsVO.ts
export interface DtsVO {
    /**
     * 
     * @type {string}
     * @memberof DtsVO
     */
    title?: string;
    /**
     * 
     * @type {string}
     * @memberof DtsVO
     */
    description?: string;
    /**
     * 
     * @type {EntityState}
     * @memberof DtsVO
     */
    state?: EntityState;
    /**
     * 
     * @type {Array<EntityState>}
     * @memberof DtsVO
     */
    allowedTransitionStates?: Array<EntityState>;
    /**
     * 
     * @type {string}
     * @memberof DtsVO
     */
    label?: string;
    /**
     * 
     * @type {string}
     * @memberof DtsVO
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof DtsVO
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof DtsVO
     */
    config?: string;
    /**
     * 
     * @type {string}
     * @memberof DtsVO
     */
    templateFk?: string;
    /**
     * 
     * @type {boolean}
     * @memberof DtsVO
     */
    debug?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof DtsVO
     */
    createdTs?: Date;
    /**
     * 
     * @type {{ [key: string]: string; }}
     * @memberof DtsVO
     */
    deploymentConfig?: { [key: string]: string; };
}