// src/openapi-client/models/EntityState.ts
/* tslint:disable */
/* eslint-disable */
/**
 * dashboard-backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const EntityState = {
    Disabled: 'DISABLED',
    Enabled: 'ENABLED',
    Archived: 'ARCHIVED',
    Editing: 'EDITING',
    Pending: 'PENDING',
    Testing: 'TESTING',
    Certified: 'CERTIFIED',
    Refused: 'REFUSED',
    Paused: 'PAUSED'
} as const;
export type EntityState = typeof EntityState[keyof typeof EntityState];


export function instanceOfEntityState(value: any): boolean {
    return Object.values(EntityState).includes(value);
}

export function EntityStateFromJSON(json: any): EntityState {
    return EntityStateFromJSONTyped(json, false);
}

export function EntityStateFromJSONTyped(json: any, ignoreDiscriminator: boolean): EntityState {
    return json as EntityState;
}

export function EntityStateToJSON(value?: EntityState | null): any {
    return value as any;
}